Refactor: wrap common/database code in a class: UploadDB
Having the database initialization done in an __init__ method instead
of at file parse time makes it easier to change configuration for
testing purposes.

However, calling Metadata() takes about 7 seconds, so we don't want
to do it a lot.  For this reason we store the engine and record_type_table_map
in class variables.