UploadedFile integer DB primary- key continued...
ValidatorHarness doesn't have access to file_id, don't try to pass it.

Rename UploadedFileChecksumEvent -> ChecksumEvent
Rename UploadedFileValidationEvent -> ValidationEvent

Add .load() to ChecksumEvent and ValidationEvent so that we may retrieve
and use existing values (file_id in particular).

IngestNotifier requires a file_id to store a notification event.  This
was previously the s3_key, but is now an integer.  This id is not passed
on to Ingest in the payload, so pass it in as an extra arg to init().