Don't notify Ingest about files that have no checksum tags.
Do a last minute check to see if the S3 object for this file still has
checksum tags.  The tags are erased if the file is overwritten, which
happens a lot as Ingest tends to upload the same file multiple times
simultaneously.  If the checksums are gone don't notify Ingest.  Some
other checksummer kicked off by the new upload will take care of that.