Switch Batch cluster instance types from "m4" to "m5"
This is required for the new validation AMI, but we may as well switch
the checksumming compute environment at the same time.

Concurrently with this change we will switch the validation cluster
ami to ami-0939b827b1bf145c9 a.k.a. ecos-with-data-vol-m5-v2

When rolling out these changes to each env, change the env's terraform.tfvars as follows:

    validation_cluster_ami_id = "ami-0939b827b1bf145c9"
    validation_cluster_instance_type = "m5"

    csum_cluster_instance_type = "m5"

To apply these terraform changes, you will need to delete the Batch job queues first
as Terraform seems to be dumb about the relationship between queues and clusters:

    tf destroy --target=module.upload-service.aws_batch_job_queue.validation_job_q
    tf destroy --target=module.upload-service.aws_batch_job_queue.csum_job_q
    make apply